Refactor context code into separate files and add more documentation
If the preferences context fails to load the preferences, it now checks
cache again for the freshest preference data.
